+++ test/testRepo/tests/testthat/test-email.R
@@ -12,3 +12,17 @@ test_that("EMAIL env var", {
     withr::with_envvar(c("EMAIL" = "bugs.bunny@acme.com"), email_address()),
     "bugs.bunny@acme.com")
 })
+
+testthat::describe("Email address", {
+  it("works", {
+    skip("Skip for test reasons")
+    mockery::stub(email_address, "system", "jambajoe@joe.joe")
+    expect_equal(email_address(), "jambajoe@joe.joe")
+  })
+
+  it("got EMAIL env var", {
+    expect_equal(
+      withr::with_envvar(c("EMAIL" = "bugs.bunny@acme.com"), email_address()),
+      "bugs.bunny@acme.com")
+  })
+})